Integer tu patří

+++ graphql_api/schema.py
@@ -241,7 +241,7 @@ class CreateProperty(Mutation, SuccessMixin):
         user_id = ID(required=True)
         usd_worth = Float(required=True)
         photo_id = String(required=True)
-        meters_squared = Float(required=True)
+        meters_squared = Int(required=True)
         property_type_id = ID(required=True)
         room_type = String(required=True)
         facility_type_ids = List(graphene.NonNull(ID), required=True)
